scripts/requirements.txt
paho-mqtt==1.5.1
influxdb==5.3.1
pandas==1.2.4
